problem c had a very simple greedy solution
how didn't I come up with solution is strange
So, simple. So, one thing is clear. It is really not
about implementation, it is about how you think of the
problem.

List out all approaches. Evaluate the correctness of each
approach and pick the one that is easier to implement
and will probably be in bound.